WT-2408 Windows error translation layer (#2819)
Build a Windows-to-POSIX/ANSI error translation layer.

Replace the read-only error mapping to WT_NOTFOUND and WT_PERM_DENIED
with EACCES and ENOENT.

Windows no longer needs its own version of __wt_strerror(), move the POSIX implementation from os_posix/os_errno.c to os_common/os_errno.c.  Rename os_win/os_errno.c to os_win/os_winerr.c to avoid a collision.

Windows now has DWORD types in prototypes, split the Windows/POSIX extern.h files. (This actually cleans up some noise, previously we had to sort the OS prototypes to remove duplicates, which wasn't trivial.)

Add the WT_EXTENSION_API.map_windows_error method to map Windows system codes to POSIX/ANSI system codes.

+++ src/support/huffman.c
@@ -813,7 +813,7 @@ __wt_huffman_decode(WT_SESSION_IMPL *session, void *huffman_arg,
 		    (bits >> (valid - max)) : (bits << (max - valid)));
 		symbol = huffman->code2symbol[pattern & mask];
 		len = huffman->codes[symbol].length;
-		valid -= len;
+		valid -= (uint8_t)len;
 
 		/*
 		 * from_len_bits is the total number of input bits, reduced by
